+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * Context for call options now have same lifetime as driver (previous - same lifetime as context for call Open function).
+* Extended metrics (fill database.sql callbacks, recognize TLI error)
+* Refactored config prefix in metrics
+* Removed excess status labels from metrics
 
 ## v3.54.2
 * Added context to some internal methods for better tracing